requirements.txt
Pillow==8.0.1
pycryptodome==3.10.1
pyDes==2.0.1
pyperclip==1.8.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.0.3
PyQt5Designer==5.14.1
PySide2==5.15.2
translate==3.5.0
typing-extensions==3.7.4.3
ujson==4.0.2

